Don't toggle WS_VISIBLE when theming is enabled
Apparently unsetting WS_VISIBLE can disrupt Direct3D presentation on the
window, so it should be avoided wherever possible. When theming is
enabled, WM_SETICON and WM_SETTEXT appear to use WM_NCUAHDRAWCAPTION to
redraw the window caption, so blocking it should be enough.

Hopefully this means the WS_VISIBLE trick is never necessary on Windows
8 and up.