Be more conservative when avoiding auto-hide appbars
Apparently ABM_GETAUTOHIDEBAR returns a valid HWND when there is an
auto-hide appbar in the specified position on any monitor, which means
if the appbar is on a secondary monitor and the window is on the primary
monitor, it will make space for an appbar that isn't there. Because of
this, the appbar avoidance logic has been changed to minimize the effect
of falsely detecting an appbar by making smaller changes to the
window's client rectangle. Now, only one side of the window is adjusted,
and it is only adjusted by 1px, so the edges of the window will be
behind any hidden appbars (like other top-level windows.)